refactor(utils): centralize getApiErrorMessage into shared apiError util
Move duplicated ApiLikeError interface and getApiErrorMessage helper (defined
locally in 10 files) into apps/web/src/utils/apiError.ts. Also export
getApiErrorStatus for status-based error branching. Replace all local casts
to ApiLikeError with the shared helper across pages, context, and components.
