+++ apps/web/src/pages/BillsPage.tsx
@@ -8,6 +8,7 @@ import {
   type BillStatusFilter,
 } from "../services/bills.service";
 import { formatCurrency } from "../utils/formatCurrency";
+import { getApiErrorMessage } from "../utils/apiError";
 
 // ─── Constants ────────────────────────────────────────────────────────────────
 
@@ -19,16 +20,6 @@ interface CategoryOption {
 }
 
 // ─── Helpers ─────────────────────────────────────────────────────────────────
-
-interface ApiLikeError {
-  response?: { data?: { message?: string } };
-  message?: string;
-}
-
-const getApiErrorMessage = (error: unknown, fallback: string): string => {
-  const e = error as ApiLikeError;
-  return e?.response?.data?.message || e?.message || fallback;
-};
 
 const formatDueDate = (dateStr: string): string => {
   if (!dateStr) return "";
